refactor: split AudioPlayer to smaller components
