+++ src/components/AudioPlayer/AudioPlayer.css
@@ -1,0 +1,5 @@
+.player.player__block {
+  width: 360px;
+  height: 60px;
+  border-radius: 8px;
+}
