+++ src/components/Cover/Cover.css
@@ -1,9 +1,3 @@
-.player.player__block {
-  width: 360px;
-  height: 60px;
-  border-radius: 8px;
-}
-
 .player.player__overlay {
   position: relative;
   z-index: 1;
@@ -22,10 +16,6 @@
   opacity: 0;
 }
 
-.player.player__overlay.player__overlay_active::after {
-  opacity: 1;
-}
-
 .player.player__cover_paused,
 .player.player__cover_playing {
   position: absolute;
@@ -39,7 +29,9 @@
 
 .player:hover .player__cover_paused,
 .player:hover .player__cover_playing,
-.player:hover .player.player__overlay::after {
+.player:hover .player.player__overlay::after,
+.player.player__overlay.player__overlay_active::after,
+.player.player__bar_visible {
   opacity: 1;
 }
 
@@ -62,10 +54,6 @@
   background-color: white;
   transform-origin: bottom;
   opacity: 0;
-}
-
-.player.player__bar_visible {
-  opacity: 1;
 }
 
 .player:hover .player.player__bar_visible {
@@ -105,37 +93,4 @@
   50% {
     transform: scaleY(8);
   }
-}
-
-.player.player__title,
-.player.player__artist {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.player.player__title {
-  letter-spacing: 0.15px;
-}
-
-.player.player__duration,
-.player.player__artist {
-  letter-spacing: 0.2px;
-  color: #818c99;
-}
-
-.player.player__info {
-  display: flex;
-  align-items: center;
-  width: 62px;
-  height: 40px;
-  padding: 0 0 0 12px;
-}
-
-.player.player__duration {
-  padding: 8px 0 8px 0;
-}
-
-.player.player__more {
-  margin: 12px 2px 12px 14px;
 }
