+++ src/components/Text/Text.css
@@ -1,0 +1,15 @@
+.player.player__title,
+.player.player__artist {
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.player.player__title {
+  letter-spacing: 0.15px;
+}
+
+.player.player__artist {
+  letter-spacing: 0.2px;
+  color: #818c99;
+}
