+++ src/components/index.js
@@ -1,0 +1,4 @@
+export { AudioPlayer } from './AudioPlayer/AudioPlayer';
+export { Cover } from './Cover/Cover';
+export { Info } from './Info/Info';
+export { Text } from './Text/Text';
